+++ tools/db_sanity_test.cc
@@ -112,7 +112,7 @@ class SanityTestSpecialComparator : public SanityTest {
     virtual int Compare(const Slice& a, const Slice& b) const override {
       return BytewiseComparator()->Compare(a, b);
     }
-    virtual double Difference(const Slice& a, const Slice& b) const override {
+    virtual int64_t Difference(const Slice& a, const Slice& b) const override {
       return BytewiseComparator()->Difference(a, b);
     }
     virtual void FindShortestSeparator(std::string* s,
